ashmem: Rewrite to improve clarity and performance
Ashmem uses a single big mutex lock for all synchronization, and even
uses it when no synchronization issues are present. The contention from
using a single lock results in all-around poor performance.

Rewrite to use fine-grained locks and atomic constructions to eliminate
the big mutex lock, thereby improving performance greatly. In places
where locks are needed for a one-time operation, we speculatively
check if locking is needed while avoiding data races. The optional name
fields are removed as well.

Note that because asma->unpinned_list never has anything added to it,
we can remove any code using it to clean up the driver a lot and
reduce synchronization requirements. This also means that
ashmem_lru_list never gets anything added to it either, so all code
using it is dead code as well, which we can remove.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Fiqri Ardyansyah <fiqri15072019@gmail.com>
Signed-off-by: dopaemon <polarisdp@gmail.com>